test($ LottoTest): 숫자 포함 확인 기능 테스트
